requirements.txt
torch
git+https://github.com/facebookresearch/detectron2.git
requests
yadisk
python-dotenv
opencv-python